Added le_wild_vars.h file containig configuration parameters. Splitted and isolated Sleep and Clock functionality
