+++ code/src/clock.h
@@ -1,0 +1,10 @@
+#ifndef _CLOCK_H
+#define _CLOCK_H
+
+int currentTime();
+int currentSeconds();
+int currentMinutes();
+int currentHours();
+
+
+#endif
