+++ code/src/le_wild_vars.h
@@ -1,0 +1,5 @@
+// SLEEP TIME THRESHOLD
+#define SLEEP_THRESHOLD_SECS 30
+
+// DISPLAY BACKLIGHT PIN
+#define BACKLIGHT_PIN 9
