+++ code/src/sleep.h
@@ -1,0 +1,9 @@
+#ifndef _SLEEP_H
+#define _SLEEP_H
+
+void wakeUpNow();
+void sleepNow();
+void checkSleepTime();
+void registerActionTime();
+
+#endif
