+++ code/src/ui_unny.h
@@ -2,7 +2,6 @@
 #define _UI_UNNY_H
 
 void setUnnyHUD(double speed);
-void checkSleepTime();
 void registerActionTime();
 void wakeUpNow();
 #endif
