+++ code/src/unny_layout.h
@@ -29,6 +29,3 @@
 //ICON SET RIGHT
 #define ICONSET_RIGHT_X 18
 #define ICONSET_RIGHT_Y 15
-
-// SLEEP TIME THRESHOLD
-#define sleepThreshold 10
